rename: BoardDomainTest 클래스 및 필드명 수정
